requirements.txt
biopython==1.83
Brotli @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/brotli-split_1699473013692/work
certifi @ file:///C:/b/abs_35d7n66oz9/croot/certifi_1707229248467/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1699472650914/work
contourpy==1.2.1
cycler==0.12.1
Cython==3.0.10
fonttools==4.51.0
idna @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/idna_1699473483982/work
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.8.4
mkl-fft @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mkl_fft_1699473528480/work
mkl-random @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mkl_random_1699473588250/work
mkl-service==2.4.0
numpy @ file:///C:/b/abs_16b2j7ad8n/croot/numpy_and_numpy_base_1704311752418/work/dist/numpy-1.26.3-cp312-cp312-win_amd64.whl#sha256=698ee7d3a0df0d67b6b5196985719e4a06dbaf5d90f3a62cb13950acc4bfb9ec
packaging==24.0
pandas==2.2.0
pillow==10.3.0
pyparsing==3.1.2
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1699473336188/work
python-dateutil==2.8.2
pytz==2024.1
requests @ file:///C:/b/abs_474vaa3x9e/croot/requests_1707355619957/work
scikit-learn==1.5.0
scipy==1.13.0
seaborn==0.13.2
setuptools==68.2.2
six==1.16.0
threadpoolctl==3.5.0
tqdm @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/tqdm_1701808178601/work
tzdata==2023.4
urllib3 @ file:///C:/b/abs_4etpfrkumr/croot/urllib3_1707770616184/work
wheel==0.41.2
win-inet-pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1699472992992/work